minimal params for summary handling db functions

+++ tldr_be/resources/migrations/20171220141137-add-summary-table.up.sql
@@ -5,4 +5,4 @@ CREATE TABLE summary
   filename VARCHAR(30),
   summary TEXT,
   votes INT DEFAULT 0,
-  doc_id BIGSERIAL);
+  doc_id INT);
